+++ lib/src/main/java/com/aeontronix/enhancedmule/tools/exchange/ExchangeClient.java
@@ -6,6 +6,7 @@ package com.aeontronix.enhancedmule.tools.exchange;
 
 import com.aeontronix.enhancedmule.tools.util.restclient.RESTClient;
 import com.aeontronix.commons.URLBuilder;
+import com.aeontronix.enhancedmule.tools.util.restclient.ResponseStream;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -19,7 +20,7 @@ public class ExchangeClient {
         this.exchangeMavenUrl = exchangeMavenUrl;
     }
 
-    public InputStream getAsset(String path) throws IOException {
+    public ResponseStream getAsset(String path) throws IOException {
         return restClient.get(new URLBuilder(exchangeMavenUrl).path("/api/v2/maven").path(path).toString()).executeReturnStream();
     }
 }